+++ src/warc2zim/cdxj_indexer/__init__.py
@@ -1,0 +1,9 @@
+from warc2zim.cdxj_indexer.bufferiter import buffering_record_iter
+from warc2zim.cdxj_indexer.main import iter_file_or_dir
+from warc2zim.cdxj_indexer.postquery import append_method_query_from_req_resp
+
+__all__ = [
+    "append_method_query_from_req_resp",
+    "buffering_record_iter",
+    "iter_file_or_dir",
+]