+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Upgrade dependencies: Python 3.13, zimscraperlib 5.1.0 and others (#434)
+- Fork cdxj_indexer codebase (#428)
 
 ## [2.2.0] - 2024-01-10